Move getBestIndexHandleForCondition
