+++ arangod/Aql/Condition.cpp
@@ -27,7 +27,7 @@
 #include "Aql/Collection.h"
 #include "Aql/ExecutionNode/EnumerateCollectionNode.h"
 #include "Aql/ExecutionPlan.h"
-#include "Aql/OptimizerUtils.h"
+#include "Aql/Optimizer/Utils/GetBestIndexHandleForFilterCondition.h"
 #include "Aql/Optimizer/Utils/GetIndexForSortCondition.h"
 #include "Aql/Quantifier.h"
 #include "Aql/Query.h"
@@ -337,7 +337,7 @@ std::pair<bool, bool> Condition::findIndexes(
       ExecutionNode::castTo<DocumentProducingNode const*>(node)
           ->canReadOwnWrites();
 
-  return aql::utils::getBestIndexHandlesForFilterCondition(
+  return aql::optimizer::getBestIndexHandlesForFilterCondition(
       trx, coll, _ast, _root, reference, sortCondition, itemsInIndex,
       node->hint(), usedIndexes, _isSorted, isAllCoveredByIndex, readOwnWrites);
 }
@@ -1132,7 +1132,7 @@ void Condition::optimize(ExecutionPlan* plan, bool multivalued) {
           ++l;
         }
       }  // cross compare sub-and-nodes
-    }  // foreach sub-and-node
+    }    // foreach sub-and-node
 
   fastForwardToNextOrItem:
     if (!retry) {
